+++ MonstroStrike/MonstroStrike/EnemyJumper.cpp
@@ -136,7 +136,7 @@ void ENEMY_JUMPER_Update(Enemy& enemy, struct Player& player)
 	}
 
 	//for gravity
-	enemy.obj.pos.y += enemy.velocity.y * AEFrameRateControllerGetFrameTime();
+	enemy.obj.pos.y += enemy.velocity.y * (f32)AEFrameRateControllerGetFrameTime();
 
 	enemy.enemyCurrent = enemy.enemyNext;
 	//main body collision box
