+++ MonstroStrike/MonstroStrike/Physics.cpp
@@ -82,5 +82,5 @@ void ApplyGravity(AEVec2 *velocity, f32 mass)
 	AEVec2 gravity = { 0.f, -9.81f }; //The acceleration aka gravity
 	AEVec2Scale(&gravityForce, &gravity, mass); //Force = Mass x Acceleration
 	AEVec2Scale(velocity, velocity, 0.98f); //Dampen, maybe add terminal velocity if needed to
-	velocity->y += gravityForce.y * AEFrameRateControllerGetFrameTime();
+	velocity->y += gravityForce.y * (f32)AEFrameRateControllerGetFrameTime();
 }
